Remove main file
Move main file to envstruct
